SVNDev/requirements.txt
amqp==2.1.4
anyjson==0.3.3
appdirs==1.4.2
arrow==0.10.0
billiard==3.5.0.2
celery==4.0.2
colorama==0.3.7
Django==1.8.17
django-celery-results==1.0.1
djangorestframework==3.5.4
httpie==0.9.9
kombu==4.0.2
packaging==16.8
Pygments==2.2.0
pymongo==3.4.0
PyMySQL==0.7.9
pyparsing==2.1.10
python-dateutil==2.6.0
pytz==2016.10
redis==2.10.5
requests==2.13.0
six==1.10.0
SQLAlchemy==1.1.6
vine==1.1.3
